+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [2018.10.03]
+
+### Added
+- `new-commit` configuration for auto committing when a new project is generated.
+- `update-commit` configuration for auto committing when a project's packages are updated.
+- `version-latest-commit` configuration for auto committing when a project's dependencies are all updated to their latest versions.
+- `version-set-commit` configuration for auto committing when a project's dependency's version is updated.
+
+### Fixed
+- JSON key used to get license info during package search
+
 ## [2018.09.08]
 
 ### Added